pyproject.toml
[project]
name = "vtu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-unfold>=0.73.1",
    "paystack>=1.5.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
]
